% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculadora.R
\name{division}
\alias{division}
\title{Division de dos numeros}
\usage{
division(x, y)
}
\arguments{
\item{x, y}{Un par de numeros cualquiera}
}
\value{
La division de los dos numeros
}
\description{
Esta funcion realiza la division de 2 numeros cualesquiera.
Los numeros pueden ser enteros o flotantes.
}
\examples{
division(4, 6)
division(3.14, 2.16)
}
