% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculadora.R
\name{multiplicacion}
\alias{multiplicacion}
\title{Multiplicacion de dos numeros}
\usage{
multiplicacion(x, y)
}
\arguments{
\item{x, y}{Un par de numeros cualquiera}
}
\value{
La multiplicacion de los dos numeros
}
\description{
Esta funcion realiza la multiplicacion de 2 numeros cualesquiera.
Los numeros pueden ser enteros o flotantes.
}
\examples{
multiplicacion(4, 6)
multiplicacion(3.14, 2.16)
}
