% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicar.R
\name{multiplicar}
\alias{multiplicar}
\title{Função para multiplicar}
\usage{
multiplicar(a, b)
}
\arguments{
\item{a}{um número}

\item{b}{outro número}
}
\description{
Função que multiplica um número pelo outro
}
