% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculadora.R
\name{resta}
\alias{resta}
\title{Resta de dos numeros}
\usage{
resta(x, y)
}
\arguments{
\item{x, y}{Un par de numeros cualquiera}
}
\value{
La resta de los dos numeros
}
\description{
Esta funcion realiza la resta de 2 numeros cualesquiera.
Los numeros pueden ser enteros o flotantes.
}
\examples{
resta(4, 6)
resta(3.14, 2.16)
}
