% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculadora.R
\name{suma}
\alias{suma}
\title{Suma de dos numeros}
\usage{
suma(x, y)
}
\arguments{
\item{x, y}{Un par de numeros cualquiera}
}
\value{
La suma de los dos numeros
}
\description{
Esta funcion realiza la suma de 2 numeros cualesquiera.
Los numeros pueden ser enteros o flotantes.
}
\examples{
suma(4, 6)
suma(3.14, 2.16)
}
